Moved internal bit representation types and functions to separate module
