+++ clash-prelude.cabal
@@ -88,6 +88,7 @@ Library
                       Clash.Annotations.Primitive
                       Clash.Annotations.BitRepresentation
                       Clash.Annotations.BitRepresentation.Deriving
+                      Clash.Annotations.BitRepresentation.Internal
                       Clash.Annotations.BitRepresentation.Util
 
                       Clash.Class.BitPack
